all-three-zones-api/requirements.txt
requests==2.31.0
beautifulsoup4==4.12.2
selenium==4.15.2
fastapi==0.104.1
uvicorn==0.24.0
python-dotenv==1.0.0
pydantic==2.5.0
aiofiles==23.2.1
webdriver-manager==4.0.1
lxml==4.9.3 